models cleanup and regenerated models.png

+++ db/migrate/20100929203845_create_packaging_options.rb
@@ -2,7 +2,7 @@ class CreatePackagingOptions < ActiveRecord::Migration
   def self.up
     create_table :packaging_options do |t|
       t.string :title
-      t.float :weight
+      t.integer :weight
       t.float :price
       t.timestamps
     end
